+++ custom_components/yahoo_jp_weather/const.py
@@ -3,6 +3,8 @@
 from datetime import timedelta
 
 DOMAIN = "yahoo_jp_weather"
+CONFIG_VERSION = 2
+USER_AGENT = "HomeAssistant YahooJapanWeather"
 CONF_ENTITY_UNIQUE_ID = "entity_unique_id"
 CONF_NAME = "name"
 DEFAULT_NAME = "江戸川区"